Implement Token Detail screen

+++ app/src/main/java/com/dreamsoftware/artcollectibles/data/api/repository/IArtMarketplaceRepository.kt
@@ -31,7 +31,7 @@ interface IArtMarketplaceRepository {
      * @param tokenId
      * @param price
      */
-    suspend fun putItemForSale(tokenId: BigInteger, price: BigInteger): BigInteger
+    suspend fun putItemForSale(tokenId: BigInteger, price: Float): BigInteger
 
     /**
      * Withdraw from sale
